Fix ICC9 compiler failure
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@11236 6f19259b-4bc3-4df7-8a09-765794883524

+++ IntelFrameworkModulePkg/Universal/BdsDxe/BootMaint/UpdatePage.c
@@ -1316,7 +1316,7 @@ UpdateSetLegacyDeviceOrderPage (
     HiiCreateOneOfOptionOpCode (
       OptionsOpCodeHandle,
       NewMenuEntry->DisplayStringToken,
-      (Index == 0) ? EFI_IFR_OPTION_DEFAULT : 0,
+      (UINT8) ((Index == 0) ? EFI_IFR_OPTION_DEFAULT : 0),
       EFI_IFR_NUMERIC_SIZE_1,
       (UINT8) ((BM_LEGACY_DEVICE_CONTEXT *) NewMenuEntry->VariableContext)->Index
       );
